implement MOVE, switch from .Rmd to .R
MOVE was a pain to implement because of problems with variable scoping. Had to use .GlobalEnv$newMSL everywhere, or a out-dated local copy got created inside apply_changes, and new taxa couldn't be seen to use as parents.

Switched to ".R" so I could use RStudio debugging. I just commented out most of the .Rmd decorations. We can switch back when we're all done.