Command-line re-write
Re-wrote the main script to parse arguments.

This involved a lot of changing the params() list and cleaning up code, re-arranging, etc.

In particular, the previous code didn't work if you didn't have .docx files, so the scanning and merging xlsx and docx lists was improved.

Caching is now setup to cache after loading all the reference info, but before scanning for proposals.
For a single 3-line xlsx file, turning on caching drops the time from about 4 seconds to about 3 seconds, so it's probably not worth the extra headaches.

All the code updating does seem to have introduced some bugs.

One odd one is that some xlsx data now seems to come in with "non-breaking space" instead of a "space", which gives the code kittens, as that doesn't count as a :space: or :blank: like it should.