IVK-225: success messages, IVK-224: QC.summary.tsv writes
Now, when each change is sucessfully processed, add a "SUCCESS" row to the errorDf. This allows parsing of output to count number of successful changes.

Also, every time we have an error, warning, success or info, write BOTH the XLSX and, now, the TSV error logs, since Don's code parses the TSV version, which was only written at the end, before, meaning it could be missing if there was a crash.