actionClean="new" & extract msl_release_num from newMSL for SQL
I had changed the "action" to "create" to make for better error
messages, but the database is expecting it to be "new", so I changed it
back.

The SQL was being generated using a params$next_msl variable that didn't exist. So created sql_new_msl and populated it from newMSL$msl_release_num.