FIX SPLIT crashes and surpress incorrect empty taxon messages
A split, with a missing/empty src or dest taxon name/lineage would crash. Fixed.

A split can create an implicit delete of a taxon. That delete was implemented AFTER the warnings about empty taxa were issued. Swapped the order so that those empty split taxa would be deleted before we scanned for erroneous empty taxa.